redefine

+++ app/src/main/java/com/example/kshitijmittal/expense/Budget.java
@@ -2,37 +2,34 @@ package com.example.kshitijmittal.expense;
 
 public class Budget {
     private String title, leftover;
-    double value;
+    int value;
     int icon;
 
-    public Budget(int icon, String title, String leftover, double value) {
+    public Budget(int icon, String title, String leftover, int value) {
         this.title = title;
         this.icon= icon;
         this.leftover = leftover;
         this.value = value;
     }
 
-    public String getTitle() {
-        return title;
+    public String getTitle() {return title;
     }
 
     public void setTitle(String name) {
         this.title = name;
     }
 
-    public String getLeftover() {
-        return leftover;
-    }
+    public String getLeftover() {return leftover;}
 
     public void setLeftover(String leftover) {
         this.leftover = leftover;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(int value) {
         this.value = value;
     }
     public int getIcon() {
